requirements.txt
python-dotenv==0.21.0
SQLAlchemy==1.3.10
aiogram==2.22.2
gino==1.0.1
asyncpg==0.27.0
emoji==2.1.0